Add solution for day 9
There's a lingering bug which is preventing me from getting the right
answers for both parts. I can get either part 1 or 2 right depending on
how I modify a conditional, but not both. I hope to return to this to
make my solution work for both parts of the question.